% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.c.R
\docType{data}
\name{data.c}
\alias{data.c}
\title{Coded dataset}
\format{
An object of class \code{data.frame} with 3744 rows and 20 columns.
}
\source{
\doi{10.1007/s40258-021-00647-3}
}
\usage{
data(data.c)
}
\description{
Coded dataset
}
\examples{
data. <- data(data.c)

}
\references{
Perez-Troncoso et al. (2021) Applied Health Economics and Health Policy 19, 757-767
(\href{https://pubmed.ncbi.nlm.nih.gov/33900567}{PubMed})
}
\keyword{datasets}
