% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\docType{data}
\name{data}
\alias{data}
\title{Survey data from the DCE in Consumers' preferences and WTP for personalised nutrition (https://doi.org/10.1007/s40258-021-00647-3)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 242 rows and 21 columns.
}
\source{
\doi{10.1007/s40258-021-00647-3}{DOI}
}
\usage{
data(survey)
}
\description{
Survey data from the DCE in Consumers' preferences and WTP for personalised nutrition (https://doi.org/10.1007/s40258-021-00647-3)
}
\examples{
data <- data(survey)

}
\references{
P?rez-Troncoso et al. (2021) Applied Health Economics and Health Policy 19, 757-767
(\href{https://pubmed.ncbi.nlm.nih.gov/33900567}{PubMed})
}
\keyword{datasets}
