% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{design}
\alias{design}
\title{Design matrix from the DCE in Consumers' preferences and WTP for personalised nutrition (https://doi.org/10.1007/s40258-021-00647-3)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 14 columns.
}
\source{
\doi{10.1007/s40258-021-00647-3}
}
\usage{
data(design)
}
\description{
Design matrix from the DCE in Consumers' preferences and WTP for personalised nutrition (https://doi.org/10.1007/s40258-021-00647-3)
}
\examples{
design <- data(design)
}
\references{
P?rez-Troncoso et al. (2021) Applied Health Economics and Health Policy 19, 757-767
(\href{https://pubmed.ncbi.nlm.nih.gov/33900567}{PubMed})
}
\keyword{datasets}
